requirements.txt
# FinTechBot Dependencies
# Pin versions to prevent supply-chain attacks

python-telegram-bot==22.6
python-dotenv==1.2.1
google-generativeai==0.8.6
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
starlette==0.52.1
